Split patch pairs into separate 2D and 3D modules
- Rename patch_pairs.py to patch_pairs_2d.py for consistency
- Keep patch_pairs_3d.py for 3D functionality
- Update all imports to use new module names
- Update documentation to reflect new module structure
- Matches project convention of separate 2D/3D files
