+++ tests/test_patch_pairs_3d.py
@@ -5,7 +5,7 @@
 import pytest
 import torch
 
-from qlty.patch_pairs import extract_overlapping_pixels_3d, extract_patch_pairs_3d
+from qlty.patch_pairs_3d import extract_overlapping_pixels_3d, extract_patch_pairs_3d
 
 
 def test_extract_patch_pairs_3d_basic():
