+++ tests/test_patch_pairs_3d_gradient.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 import torch.nn as nn
 
-from qlty.patch_pairs import extract_overlapping_pixels_3d, extract_patch_pairs_3d
+from qlty.patch_pairs_3d import extract_overlapping_pixels_3d, extract_patch_pairs_3d
 
 
 def test_kernel_optimization_3d_with_overlapping_pixels():
